move plot_alpha to results_utils
