+++ tests/test_per_sample.py
@@ -2,7 +2,6 @@ import os
 import shutil
 from unittest import TestCase, main
 
-import numpy as np
 import pandas as pd
 import numpy.testing as npt
 
@@ -209,41 +208,6 @@ class PerSampleTests(TestCase):
         exp = {'foo': "Cannot stage template."}
         obs = agps.stage_per_sample_specific_statics(opts, ids)
         self.assertEqual(obs, exp)
-
-    def test_plot_alpha(self):
-        map_ = pd.DataFrame(
-            data=np.array([
-                ['skin', '1990', 'female', 'Verity', 'US', 12.5],
-                ['fecal', '1990', 'female', 'Verity', 'US', 8.6],
-                ['fecal', '1987', 'male', 'Alex', 'US', 7.9],
-                ['fecal', '1993', 'female', 'Annie', 'US', 7.5],
-                ['skin', '1989', 'male', 'Dominic', 'UK', 14.0],
-                ['fecal', '1986', 'female', 'Sarah', 'US', 15.0],
-                ['oral', '1988', 'female', 'Shelby', 'AUS', 4.2],
-                ]),
-            index=['VeP0', 'VeP1', 'AxP0', 'AnP0', 'DoD0', 'SaZ0', 'ShT0'],
-            columns=['SIMPLE_BODY_SITE', 'BIRTH_YEAR', 'SEX',
-                     'HOST_SUBJECT_ID', 'NATIONALITY', 'alpha'],
-            )
-        sample = 'VeP0'
-        kgroup = 'skin'
-        kgalpha = pd.Series([12.5, 14.0],
-                            index=['VeP0', 'DoD0'],
-                            name='alpha',
-                            dtype=object)
-        ksalpha = 12.5
-        kxlabel = 'alphadiversity'
-
-        (tgroup, tgalpha, tsalpha, txlabel) = \
-            agps._plot_alpha(sample=sample,
-                             alpha_map=map_,
-                             alpha_field='alpha',
-                             debug=True)
-
-        self.assertEqual(tgroup, kgroup)
-        npt.assert_array_equal(tgalpha.values, kgalpha.values)
-        self.assertEqual(tsalpha, ksalpha)
-        self.assertEqual(txlabel, kxlabel)
 
     def test_alpha_plot_metric_error_field_error(self):
         opts = {'collapsed': {
